Developpement

+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,18 +1,26 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
 {
-    public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantsRepository restaurantsRepository)
+    public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, 
+        IRestaurantsRepository restaurantsRepository,
+        IRestaurantAuthorizationService restaurantAuthorizationService)
         : IRequestHandler<DeleteRestaurantCommand>
     {
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
             logger.LogInformation($"Delete restaurant with id : {request.Id}");
             var restaurant = await restaurantsRepository.GetByIdAsync(request.Id) ?? throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+            if (!restaurantAuthorizationService.Authorize(restaurant, RessourceOperation.Delete))
+            {
+                throw new ForbidException();
+            }
             await restaurantsRepository.Delete(restaurant);
         }
     }
